refactor(direct-control): extract unit move preview source
Move the unit move preview embedded App UI source into src/play/ready/move-preview.ts while keeping the public wrapper and command builder in the facade. Keep the move as source ownership only; this does not claim live-game runtime proof.

Record the ready atom status in the support modularization OpenSpec workstream, including the remaining ready-unit and ready-city source owners.

Validation: bun run --cwd packages/civ7-direct-control test -- unit-move-preview.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
